Web Application/requirements.txt
absl-py==2.1.0
accelerate==1.0.1
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2
bitsandbytes==0.44.1
brotli-python==1.1.0
datasets==3.0.1
flask==3.0.3
fpdf==1.7.2
fsspec==2024.6.1
h5py==3.11.0
huggingface-hub==0.25.2
jax==0.4.30
jaxlib==0.4.30
keras==2.12.0
numpy==1.24.3  
scipy==1.11.3  
tensorflow==2.13.0 
torch==2.1.0  
transformers==4.35.0
whisper==1.2.0  
